AutomatedFormFilling/requirements.txt
streamlit
openai
requests
fpdf
google-cloud-speech
sounddevice